Split up tasks into an app and added extra fields
