+++ ocdaction/ocdaction/settings/__init__.py
@@ -31,9 +31,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'profiles',
     'custom_user',
-    'core'
+    'profiles',
+    'core',
+    'tasks'
 
 ]
