+++ ocdaction/profiles/urls.py
@@ -1,3 +1,7 @@
+"""
+    ocdaction URL Configuration
+    profile/user specific urls for auth workflows
+"""
 from django.conf.urls import include, url
 from django.contrib.auth import views as auth_views
